requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
nltk==3.8.1
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.0.9
wordcloud==1.9.3
matplotlib==3.8.4
joblib==1.3.2
